requirements.txt
streamlit==1.56.0
langchain==1.2.15
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-text-splitters==1.1.2
langchain-google-genai==4.2.2
google-generativeai==0.8.6
google-genai==2.3.0
chromadb==1.5.8
PyMuPDF==1.27.2.2
docx2txt==0.9
python-dotenv==1.2.2
protobuf==6.33.6
pillow==12.2.0
pandas==2.3.3
matplotlib==3.10.9
openpyxl==3.1.5
